requirements.txt
# Core dependencies
Pillow>=9.0.0  # Image processing library
datasets>=2.0.0  # HuggingFace datasets library
tqdm>=4.62.0  # Progress bar library
piexif>=1.1.3  # For handling EXIF metadata in images

# Database and storage
# sqlite3 is included in Python standard library

# Optional dependencies
numpy>=1.20.0  # For numerical operations
requests>=2.27.0  # For HTTP requests
huggingface-hub>=0.10.0  # For HuggingFace dataset access
psutil>=5.9.0  # For memory monitoring

# Development dependencies
pytest>=7.0.0  # For unit testing
black>=22.0.0  # For code formatting
flake8>=4.0.0  # For linting
